Added numbering if multiple file in a day

+++ main.py
@@ -2,6 +2,7 @@ from pydrive2.drive import GoogleDrive
 from pydrive2.auth import GoogleAuth
 from datetime import date
 import os
+import time
 
 gauth = GoogleAuth()
 
@@ -26,24 +27,19 @@ gauth.SaveCredentialsFile("mycreds.txt")
 drive = GoogleDrive(gauth)
 
 path = r"/home/pi/Documents/vods/vid"
-
+i = 1
 for file in os.listdir(path):
-  newFile = drive.CreateFile({'title': f"{date.today()}--wubby", 'parents': [{'id': '1n1eeqQH4mr1J_gv_-d_cp1sqAj0nYNl_'}]})
+  newFile = drive.CreateFile({'title': f"{date.today()}--wubby-{i}", 'parents': [{'id': '1n1eeqQH4mr1J_gv_-d_cp1sqAj0nYNl_'}]})
   newFile.SetContentFile(os.path.join(path, file))
-  try: 
-    newFile.Upload()
-    print(f" {file} has been uploaded to Google Drive")
-    os.remove(os.path.join(path, file))
-    print(f" Deleted: {file}")
-  except:
-    f" Failed to upload {file} to Google Drive"
+  newFile.Upload()
+  print(f" {file} has been uploaded to Google Drive")
+  i+=1
   f = None
-
-# for file in os.listdir(path):
-#   age = os.stat(os.path.join(path, file)).st_mtime
-#   if age < time.time() - 86400*14:
-#     print(f" Deleted: {file}")
-#     os.remove(os.path.join(path, file))
-#   else:
-#     print(f" {file} is too young to be deleted!")
-#   st = None
+  
+for file in os.listdir(path):
+  age = os.stat(os.path.join(path, file)).st_mtime
+  if age < time.time() - 86400:
+    print(f" Deleted: {file}")
+    os.remove(os.path.join(path, file))
+  else:
+    print(f" {file} is too young to be deleted")
